# Port native_compute to OCaml 5.