This pull request is not meant to be applied, since it is way too fragile as it is. It only exists because @gasche needed to experiment with `native_compute` on OCaml 5, so I might as well make it public.

The first commit is the actual port. It changes accumulators so that they no longer have tag 0 but tag 247 (as any closure should under OCaml 5). This makes `native_compute` 160% slower, compared to OCaml 4 with tag-0 accumulators.

The second commit is a 64-bit, little-endian hack to test what the actual cost of `Obj.tag` is. The slowdown is reduced to just 10%. In other words, 60% of the execution time of `native_compute` is actually wasted inside `Obj.tag` alone. Indeed, as surprising as it might be, `Obj.tag` is not a builtin construct but a dumb C function.

(@gasche The timing difference for the hack is because there was an occurrence of `Obj.tag` that I mistakenly thought had negligible impact. Now that I have hopefully removed all the impactful instances of `Obj.tag`, the performances of OCaml 5 are almost on par with OCaml 4.)