@ppedrot 2025-03-26
Where do your 160% ~ 10% values come from? I wrote basically the same patch at some point and on some microbenchmarks I got way worse results.