@ppedrot 2025-03-26
As per #14048 (more precisely [this comment](https://github.com/coq/coq/pull/14048#issuecomment-815007067)) it seems that UnsaturatedSolinasHeuristics/Tests.v in fiat-crypto is a good perf test for this kind of change.