@silene 2025-03-26
I am using my own `native_compute`-based proofs as a benchmark. But here is a toy example (4M terms of Collatz' conjecture) that exhibits the kind of performances I am experiencing:

- OCaml 4.13.1 => 4.9s
- OCaml 4.13.1 + tag-247 => 13.6s
- OCaml 5.1.1 + tag-247 => 12.7s
- OCaml 5.1.1 + tag-247 + unsafe_get => 5.5s

```coq
From Coq Require Import Uint63.

Fixpoint run n v :=
  if leb v 1 then v else
  match n with
  | O => if is_even v then lsr v 1 else lsr (3 * v + 1) 1
  | S n => run n (run n v)
  end.

Fixpoint run_many n (vl : int * list int) :=
  match n with
  | O =>
    let (v, l) := vl in
    let l := if eqb (run 20 v) 1 then l else List.cons v l in
    (add v 1, l)
  | S n => run_many n (run_many n vl)
  end.

Time Eval native_compute in run_many 22 (1%int63, nil).
```