@silene 2025-03-27
In the following case (`rennertypeD51.v`, not public), I get timings closer to the 5x slowdown you experienced at the time.

- OCaml 4.13.1 => 61.8s
- OCaml 5.1.1 + tag-247 => 320.s
- OCaml 5.1.1 + tag-247 + unsafe_get => 88.2s