@silene 2025-03-27
No, I don't think we have anything interesting that uses `native_compute` in the bench. (Or do we?)