@SkySkimmer 2025-03-27
coq-neural-net-interp-computed-lite is a native compute user