@silene 2025-03-27
Hmm... It uses `#[native_compile=no]` all over the place. And from a cursory glance, I cannot find a single occurrence of `native_compute` or `native_cast_no_check` that is not commented out. Is the version in the CI really using `native_compute`?