@SkySkimmer 2025-03-27
You see https://github.com/coq/coq/pull/19569#issuecomment-2360320021
For instance https://github.com/JasonGross/neural-net-coq-interp/blob/7e0c38173eb9147bde15e17ac43583df96289465/theories/MaxOfTwoNumbers/Computed/AllLogits.v#L8
~~~coq
Time #[native_compile=no] Local Definition all_tokens_logits_concrete_value := native_compute logits_all_tokens_concrete_opt.
~~~

native_compile=0 means the resulting value is not compiled, but it is still computed using native_compute