@silene 2025-03-27
Sure, but that file does not seem to be compiled according to the CI log: https://gitlab.inria.fr/coq/coq/-/jobs/5516692/raw

Am I missing something?