@silene 2025-03-27
Oh! I guess what you mean that `native_compute` is not tested in the CI, unless the bench is run. Is that it?