@SkySkimmer 2025-03-27
The bench isn't the same as CI. CI runs some custom scripts such as https://github.com/coq/coq/blob/master/dev/ci/scripts/ci-neural_net_interp.sh
The bench runs opam packages such as https://github.com/coq/opam/blob/master/extra-dev/packages/coq-neural-net-interp-computed-lite/coq-neural-net-interp-computed-lite.dev/opam