@silene 2025-03-27
`dev/bench/bench.sh` contains `coq-neural-net-interp-computed-lite`; `coq-neural-net-interp-computed-lite.dev/opam` contains `build: [make "-j%{jobs}%" "computed-lite"]`; `neural-net-coq-interp/Makefile.coq.local` contains `computed-lite: theories/MaxOfTwoNumbersSimpler/Computed/AllLogits.vo`. I do not see what could cause the other files to be compiled, especially the not "Simpler" version. The bench you cite seems to contain more tests, but it was a custom bench, not the normal one, I think.