@JasonGross 2025-03-27
> I do not see what could cause the other files to be compiled, especially the not "Simpler" version.

Note that all uses of native-compute in nnci are essentially the same: linear algebra done on primitive arrays of primitive floats indexed by primitive ints, with loops run on well-founded recursion, iirc (I'm interested in knowing if there are design patterns for this that are not 40,000x slower than pytorch).  And they all produce enormous arrays of arrays of floats.  In contrast, the Fiat Crypto examples do moderate amount of computation (with Z, IIRC), and produce relatively small outputs.