@SkySkimmer 2025-03-27
>I do not see what could cause the other files to be compiled, especially the not "Simpler" version.

AllLogits has dependencies