@Zimmi48 2025-03-27
The bench seems to have failed because `coq-perennial` is missing. Any idea why? (I was looking to see if the failure was related to the org and repo renaming, but it doesn't look like it.)