@SkySkimmer 2025-03-27
Failed because rocq-native declares itself to be incompatible with ocaml 5